Add in virtual host for findthemakers.com.
Also split out the routing in server.js to make it more readable.
