+++ package.json
@@ -23,7 +23,8 @@
     "express": "^4.17.1",
     "express-handlebars": "^4.0.3",
     "handlebars-helpers": "^0.10.0",
-    "http-proxy-middleware": "^1.0.3"
+    "http-proxy-middleware": "^1.0.3",
+    "vhost": "^3.0.2"
   },
   "devDependencies": {
     "@babel/cli": "^7.8.4",
